==> notebooks/2-data-analysis-hate.ipynb <==
# Data For Hate Methodology
Each header in this notebook corresponds to a headers in the methodology.

import os
import pandas as pd
from tldextract import extract
from utils import value_counts

# inputs
fn_policy = '../data/output/placements_api_keyword_status/policy.csv'
fn_hate = '../data/output/placements_api_keyword_status/hate.csv'
fn_adhoc = '../data/output/placements_api_keyword_status/adhoc.csv'
fn_basewords = '../data/output/placements_api_keyword_status/basewords.csv'
fn_youtube_videos = '../data/output/placements_api_suggestions/videos_for_hate_terms.csv'
fn_youtube_channels = '../data/output/placements_api_suggestions/channels_for_hate_terms.csv'

# outputs
table_dir = '../data/output/tables/hate'
fn_table1 = '../data/output/tables/hate/table1.csv'
fn_table2 = '../data/output/tables/hate/table2.csv'
fn_table3a = '../data/output/tables/hate/table3a.csv'
fn_table3b = '../data/output/tables/hate/table3b.csv'
fn_table3a_full = '../data/output/tables/hate/table3a_full.csv'
fn_table3b_full = '../data/output/tables/hate/table3b_full.csv'
os.makedirs(table_dir, exist_ok=True)

# columns we'll display in the notebook
display_cols = ['search_term', 'status']

df_policy = pd.read_csv(fn_policy)
df_hate = pd.read_csv(fn_hate)
df_basewords = pd.read_csv(fn_basewords)
df_adhoc = pd.read_csv(fn_adhoc)

## Google Ads API for Ad Placements

# how many terms that YouTube saya are not advertiser-friendly are "blocked"?
value_counts(df_policy, col='status')

Examples of `Full`, `Blocked`, and `Partial Block` terms:

(df_policy[df_policy.status == 'Full']
    .sample(3, random_state=303)
    .search_term.tolist())

(df_policy[df_policy.status == 'Blocked']
    .sample(3, random_state=303)
    .search_term.tolist())

(df_policy[df_policy.status == 'Partial Block']
    .sample(3, random_state=303)
    .search_term.tolist())

## Whats on the blocklist?

col2hate_table_col = {
    'search_term': "Search Term", 
    'status': "Status for Ad Placement",
    'n_youtube_videos': "N YouTube Videos",
    'n_youtube_channels': "N YouTube Channels",
    "background_info_link": "Background Info Link"
}

domain2name = {
    'adl.org': 'ADL',
    'splcenter.org': 'SPLC',
    'washingtonpost.com': 'Washington Post',
    'vice.com': 'Vice',
    'timesofisrael.com': 'Times of Israel',
    'isdglobal.org': 'ISD',
    'theguardian.com': 'The Guardian',
    'npr.org': 'NPR',
    'washingtonmonthly.com': 'Washington Monthly',
    'knowyourmeme.com': 'Know Your Meme',
    'newsweek.com': 'Newsweek',
    'cjr.org': 'CJR',
    'oup.com': 'OUP',
    'buzzfeednews.com': 'BuzzFeed News',
    'thedailybeast.com': 'The Daily Beast',
    'psmag.com': 'Pacific Standard',
    'haaretz.com': 'Haaretz',
    'ushmm.org': 'USHMM',
    'gmu.edu': 'George Mason U.',
    'advocate.com': 'The Advocate',
    'newyorker.com': 'The New Yorker'
}

def clean_domain(link):
    tsd, td, tsu = extract(link) # prints abc, hostname, com
    url = td + '.' + tsu # will prints as hostname.com
    return domain2name.get(url, url)

hate_table = df_hate[[c for c in col2hate_table_col.keys()]]
hate_table.columns = col2hate_table_col.values()

# get the info source name
hate_table['Background Info'] = hate_table['Background Info Link'].apply(clean_domain)

hate_table.to_csv(fn_table1, index=False)
hate_table.head(5)

# what are the statuses of un-altered hate terms?
value_counts(df_hate, col='status')

# what are the statuses from multi-word terms with spaces removed?
df_hate_phrases = df_hate[df_hate.search_term.str.split(' ').str.len() > 1]
value_counts(df_hate_phrases,
             col='status_no_spaces')

# what is still blocked after removing spaces?
still_blocked = df_hate[
    (df_hate.search_term.str.split(' ').str.len() > 1) &
    (df_hate.status_no_spaces == "Blocked")
].search_term.tolist()

still_blocked

# among those three terms, which base words are blocked?
unique_basewords = {
    word for sent in still_blocked for word in sent.split(' ')
}

df_basewords[df_basewords.search_term.isin(unique_basewords)].append(
    df_hate[df_hate.search_term == 'white pride']
)[display_cols].set_index('search_term')

# white nationalist vs -ists and -ism
(df_hate[df_hate.search_term.str.contains('white national')]
    .append(df_adhoc[df_adhoc.search_term.str.contains('white national')])
    .sort_values(by='status')[display_cols]
    .set_index('search_term'))

# "terrorists" vs "terrorist"
(df_basewords[df_basewords.search_term.str.contains('terrorist')]
    .append(df_policy[df_policy.search_term.str.contains('terrorist')])
    .append(df_adhoc[df_adhoc.search_term.str.contains('terrorist')])
    .sort_values(by='status')[display_cols]
    .set_index('search_term'))

# multi-word blocked terms, which return full responses after spaces are removed.
bandaids = df_hate[
    (df_hate.status == 'Blocked') &
    (df_hate.status_no_spaces == "Full")
][['search_term']]

bandaids['full response'] = bandaids['search_term'].str.replace(' ', '')

bandaids.columns = ['blocked response', 'full response']
bandaids.sort_values(by='blocked response', inplace=True, ascending=False)
bandaids.to_csv(fn_table2, index=False)
bandaids

## Hate phrases vs hate content

# how many videos are related to "white power"?
df_hate[df_hate.search_term == "white power"].n_youtube_videos

df_hate_video = pd.read_csv(fn_youtube_videos)
df_hate_channels = pd.read_csv(fn_youtube_channels)

What are the most suggested channels?

# top channels from suggested videos
channels_from_video = df_hate_video.youtube_video_channel.value_counts()
channels_from_video.to_csv(fn_table3a_full)

# this is what is displayed in the method
channels_from_video = channels_from_video[channels_from_video >= 8]
channels_from_video.to_csv(fn_table3a)
channels_from_video

# top channels from suggested channels
channels_from_channels = df_hate_channels.youtube_channel_name.value_counts()
channels_from_channels.to_csv(fn_table3b_full)

# this is what is displayed in the method
channels_from_channels = channels_from_channels[channels_from_channels>= 3]
channels_from_channels.to_csv(fn_table3b)
channels_from_channels
