==> notebooks/4-data-analysis-social-justice.ipynb <==
# Data For Social Justice Methodology
Each header in this notebook corresponds to a headers in the methodology.

import os
import pandas as pd
from utils import value_counts

# inputs
fn_policy = '../data/output/placements_api_keyword_status/policy.csv'
fn_hate = '../data/output/placements_api_keyword_status/hate.csv'
fn_adhoc = '../data/output/placements_api_keyword_status/adhoc.csv'
fn_basewords = '../data/output/placements_api_keyword_status/basewords.csv'
fn_social_justice = '../data/output/placements_api_keyword_status/social_justice.csv'
fn_youtube_videos = '../data/output/placements_api_suggestions/videos_for_social_justice_terms.csv'
fn_youtube_channels = '../data/output/placements_api_suggestions/channels_for_social_justice_terms.csv'

# outputs
table_dir = '../data/output/tables/socialjustice'
fn_table1 = '../data/output/tables/socialjustice/table1.csv'
fn_table2 = '../data/output/tables/socialjustice/table2.csv'
fn_table3 = '../data/output/tables/socialjustice/table3.csv'
os.makedirs(table_dir, exist_ok=True)

# columns we'll display in the notebook
display_cols = ['search_term', 'status']
display_cols_bandaids = ['search_term', 'status', 'status_no_spaces']

# how we're going to sort results
sort_cols = ['status', 'search_term']
sort_cols_bandaids = ['status', 'status_no_spaces', 'search_term']

df_soc_just = pd.read_csv(fn_social_justice)
df_adhoc = pd.read_csv(fn_adhoc)
df_policy = pd.read_csv(fn_policy)
df_basewords = pd.read_csv(fn_basewords)
df_hate = pd.read_csv(fn_hate)

## What is Blocked?

# create order of statuses for sorting
status_order = ["Blocked", "Partial Block", "Empty", "Full"]
for col in ['status', 'status_no_spaces']:
    df_soc_just[col] = pd.Categorical(df_soc_just[col], status_order)

col2social_justice_table_col = {
    'search_term': "Search Term", 
    'status': "Statusfor Ad Placement",
    'status_no_spaces': "Status with Spaces Removed",
    'n_youtube_videos': "N YouTube Videos",
    'n_youtube_channels': "N Channels Videos",
}

social_justice_table = df_soc_just[[c for c in col2social_justice_table_col.keys()]]
social_justice_table = social_justice_table.sort_values(
    by=['status', 'status_no_spaces', 'search_term'], 
    ascending=(True, False, True)
)
social_justice_table.columns = col2social_justice_table_col.values()
social_justice_table.to_csv(fn_table1, index=False)
social_justice_table.head(10).reset_index(drop=True)

social_justice_table

value_counts(df_soc_just, col='status')

# what terms have a partial block?
df_soc_just[df_soc_just.status == 'Partial Block'].search_term.tolist()

## White power vs Black power

whose_power = df_soc_just[df_soc_just.search_term == 'black power'].append(
    df_hate[df_hate.search_term == 'white power']
)[display_cols]

whose_power.set_index('search_term')

whose_lives_matter_to_yt = (df_soc_just[df_soc_just.search_term.str.contains('matter')]
    .append(df_hate[df_hate.search_term == 'white lives matter'])
    .append(df_adhoc[df_adhoc.search_term.str.contains('matter')]
)[display_cols].sort_values(by=sort_cols))

whose_lives_matter_to_yt.set_index('search_term')

Create a table for the methodology

blocked = whose_power[whose_power.status == 'Blocked'].search_term.tolist() + \
    whose_lives_matter_to_yt[whose_lives_matter_to_yt.status == 'Blocked'].search_term.tolist()
    
full = whose_power[whose_power.status == 'Full'].search_term.tolist() + \
    [', '.join(whose_lives_matter_to_yt[whose_lives_matter_to_yt.status == 'Full'].search_term.tolist())]

table2 = pd.DataFrame({
    "Blocked response": blocked,
    "Full response": full
})
table2

# save the table
table2.to_csv(fn_table2, index=False,  escapechar='"')

The next three cells look at terms containing "muslim":

# How are religions treated as basewords...
religions = ['buddhist', 'jewish', 'christian', 'muslim']
relgion = df_basewords[df_basewords.search_term.isin(religions)][display_cols]
relgion.set_index('search_term')

# How are terms that contain "muslim" treated?
df_soc_just[
    df_soc_just.search_term.str.contains('muslim')
][display_cols].set_index('search_term')

# Are there religions treated the same way?
fashion = df_soc_just[df_soc_just.search_term.str.contains('muslim fashion')].append(
    df_adhoc[df_adhoc.search_term.str.contains('fashion')]
)[display_cols]

fashion.set_index('search_term')

df_soc_just[df_soc_just.search_term.str.contains('muslim parenting')].append(
    df_adhoc[df_adhoc.search_term.str.contains('parenting')]
)[display_cols].sort_values(by='status').set_index('search_term')

let's format this into a table for the methodology

# this is to combine "muslim" and "muslim fashion" as one entry
muslim_fashion = {
   "muslim" : "muslim, muslim fashion"   
}
blocked_religion = relgion[relgion.status == 'Blocked'].search_term.tolist()
blocked_religion = [muslim_fashion.get(term, term) for term in blocked_religion]

full_religion = fashion[fashion.status == 'Full'].search_term.tolist()
full_religion.append(None)

table3 = pd.DataFrame({
    "Blocked response": blocked_religion,
    "Full response": full_religion
})
table3

table3.to_csv(fn_table3, index=False, escapechar='"')

# What's the status of blocked terms with spaces removed?
value_counts(df_soc_just[df_soc_just.status == 'Blocked'], 
             col='status_no_spaces')

# What terms are still blocked with spaces removed?
(df_soc_just[df_soc_just.status == 'Blocked'][display_cols_bandaids ]
    .sort_values(by=sort_cols, ascending=False)
    .set_index('search_term'))

# anti-muslim terms from the hate list:
anti_muslim = ['civilization jihad', 'white sharia', 'radical islamic terror']
(df_hate[df_hate.search_term.isin(anti_muslim)][display_cols]
    .set_index('search_term'))

## Band-Aids
"covid" and "sex" are blocked as words and substrings.

(df_policy[df_policy.search_term.str.contains('covid-19')]
    .append(df_basewords[df_basewords.search_term.str.contains('covid-19')])
    .sort_values(sort_cols_bandaids)[display_cols_bandaids]
    .set_index('search_term'))

(df_policy[df_policy.search_term.str.contains('sex')]
    .append(df_basewords[df_basewords.search_term.str.contains('sex')])
    .sort_values(sort_cols_bandaids)[display_cols_bandaids]
    .set_index('search_term'))

What channels are suggested?

df_video = pd.read_csv(fn_youtube_videos)
df_channel = pd.read_csv(fn_youtube_channels)

channels_from_videos = df_video.youtube_video_channel.value_counts()
channels_from_channels = df_channel.youtube_channel_name.value_counts()

channels_from_videos.head(20)

channels_from_channels.head(20)
